@@ -10,32 +10,28 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.10.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  firebase_auth: ^0.16.1
-  google_sign_in: ^4.0.7
-  flutter_facebook_login: ^2.0.1
-  provider: ^4.0.1
-  cloud_firestore: ^0.13.7
-  intl: ^0.16.0
-  rxdart: ^0.23.1
+  cloud_firestore: ^0.14.2
+  cupertino_icons: ^1.0.0
+  firebase_auth: ^0.18.2
+  firebase_core: ^0.5.1
+  flutter_login_facebook: ^0.3.0
+  google_sign_in: ^4.5.6
+  intl: ^0.16.1
+  provider: ^4.3.2+2
+  rxdart: ^0.24.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   mockito: ^4.1.1
 
-
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
-
 # The following section is specific to Flutter.
 flutter:
 
@@ -49,15 +45,12 @@ flutter:
   #  - images/a_dot_burr.jpeg
   #  - images/a_dot_ham.jpeg
   assets:
-   - images/google-logo.png
-   - images/facebook-logo.png
-
+    - images/google-logo.png
+    - images/facebook-logo.png
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.io/assets-and-images/#resolution-aware.
-
   # For details regarding adding assets from package dependencies, see
   # https://flutter.io/assets-and-images/#from-packages
-
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
   # "family" key with the font family name, and a "fonts" key with a
