@@ -45,10 +45,7 @@ class CupertinoHomeScaffold extends StatelessWidget {
         itemData.icon,
         color: color,
       ),
-      title: Text(
-        itemData.title,
-        style: TextStyle(color: color),
-      ),
+      label: itemData.title,
     );
   }
 }
