@@ -1,13 +1,13 @@
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 import 'package:time_tracker_flutter_course/app/home/job_entries/job_entries_page.dart';
 import 'package:time_tracker_flutter_course/app/home/jobs/edit_job_page.dart';
 import 'package:time_tracker_flutter_course/app/home/jobs/job_list_tile.dart';
 import 'package:time_tracker_flutter_course/app/home/jobs/list_items_builder.dart';
 import 'package:time_tracker_flutter_course/app/home/models/job.dart';
-import 'package:time_tracker_flutter_course/common_widgets/platform_exception_alert_dialog.dart';
+import 'package:time_tracker_flutter_course/common_widgets/show_exception_alert_dialog.dart';
 import 'package:time_tracker_flutter_course/services/database.dart';
 
 class JobsPage extends StatelessWidget {
@@ -15,11 +15,12 @@ class JobsPage extends StatelessWidget {
     try {
       final database = Provider.of<Database>(context, listen: false);
       await database.deleteJob(job);
-    } on PlatformException catch (e) {
-      PlatformExceptionAlertDialog(
+    } on FirebaseException catch (e) {
+      showExceptionAlertDialog(
+        context,
         title: 'Operation failed',
         exception: e,
-      ).show(context);
+      );
     }
   }
 
