@@ -1,3 +1,4 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:time_tracker_flutter_course/app/home/home_page.dart';
@@ -6,11 +7,14 @@ import 'package:time_tracker_flutter_course/services/auth.dart';
 import 'package:time_tracker_flutter_course/services/database.dart';
 
 class LandingPage extends StatelessWidget {
+  const LandingPage({Key key, this.databaseBuilder}) : super(key: key);
+  final Database Function(String) databaseBuilder;
+
   @override
   Widget build(BuildContext context) {
     final auth = Provider.of<AuthBase>(context, listen: false);
     return StreamBuilder<User>(
-        stream: auth.onAuthStateChanged,
+        stream: auth.authStateChanges(),
         builder: (context, snapshot) {
           if (snapshot.connectionState == ConnectionState.active) {
             User user = snapshot.data;
@@ -20,7 +24,7 @@ class LandingPage extends StatelessWidget {
             return Provider<User>.value(
               value: user,
               child: Provider<Database>(
-                create: (_) => FirestoreDatabase(uid: user.uid),
+                create: (_) => databaseBuilder(user.uid),
                 child: HomePage(),
               ),
             );
