@@ -1,10 +1,11 @@
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 import 'package:time_tracker_flutter_course/app/sign_in/email_sign_in_bloc.dart';
 import 'package:time_tracker_flutter_course/app/sign_in/email_sign_in_model.dart';
 import 'package:time_tracker_flutter_course/common_widgets/form_submit_button.dart';
-import 'package:time_tracker_flutter_course/common_widgets/platform_exception_alert_dialog.dart';
+import 'package:time_tracker_flutter_course/common_widgets/show_exception_alert_dialog.dart';
 import 'package:time_tracker_flutter_course/services/auth.dart';
 
 class EmailSignInFormBlocBased extends StatefulWidget {
@@ -12,7 +13,7 @@ class EmailSignInFormBlocBased extends StatefulWidget {
   final EmailSignInBloc bloc;
 
   static Widget create(BuildContext context) {
-    final AuthBase auth = Provider.of<AuthBase>(context, listen: false);
+    final auth = Provider.of<AuthBase>(context, listen: false);
     return Provider<EmailSignInBloc>(
       create: (context) => EmailSignInBloc(auth: auth),
       child: Consumer<EmailSignInBloc>(
@@ -46,11 +47,12 @@ class _EmailSignInFormBlocBasedState extends State<EmailSignInFormBlocBased> {
     try {
       await widget.bloc.submit();
       Navigator.of(context).pop();
-    } on PlatformException catch (e) {
-      PlatformExceptionAlertDialog(
+    } on FirebaseException catch (e) {
+      showExceptionAlertDialog(
+        context,
         title: 'Sign in failed',
         exception: e,
-      ).show(context);
+      );
     }
   }
 
