@@ -1,9 +1,10 @@
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 import 'package:time_tracker_flutter_course/app/sign_in/email_sign_in_change_model.dart';
 import 'package:time_tracker_flutter_course/common_widgets/form_submit_button.dart';
-import 'package:time_tracker_flutter_course/common_widgets/platform_exception_alert_dialog.dart';
+import 'package:time_tracker_flutter_course/common_widgets/show_exception_alert_dialog.dart';
 import 'package:time_tracker_flutter_course/services/auth.dart';
 
 class EmailSignInFormChangeNotifier extends StatefulWidget {
@@ -11,7 +12,7 @@ class EmailSignInFormChangeNotifier extends StatefulWidget {
   final EmailSignInChangeModel model;
 
   static Widget create(BuildContext context) {
-    final AuthBase auth = Provider.of<AuthBase>(context, listen: false);
+    final auth = Provider.of<AuthBase>(context, listen: false);
     return ChangeNotifierProvider<EmailSignInChangeModel>(
       create: (context) => EmailSignInChangeModel(auth: auth),
       child: Consumer<EmailSignInChangeModel>(
@@ -51,11 +52,12 @@ class _EmailSignInFormChangeNotifierState
     try {
       await model.submit();
       Navigator.of(context).pop();
-    } on PlatformException catch (e) {
-      PlatformExceptionAlertDialog(
+    } on FirebaseException catch (e) {
+      showExceptionAlertDialog(
+        context,
         title: 'Sign in failed',
         exception: e,
-      ).show(context);
+      );
     }
   }
 
