@@ -1,23 +1,10 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_facebook_login/flutter_facebook_login.dart';
+import 'package:flutter_login_facebook/flutter_login_facebook.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 
-class User {
-  User({
-    @required this.uid,
-    this.photoUrl,
-    this.displayName,
-  });
-  final String uid;
-  final String photoUrl;
-  final String displayName;
-}
-
 abstract class AuthBase {
-  Stream<User> get onAuthStateChanged;
-  Future<User> currentUser();
+  Stream<User> authStateChanges();
+  User get currentUser;
   Future<User> signInAnonymously();
   Future<User> signInWithEmailAndPassword(String email, String password);
   Future<User> createUserWithEmailAndPassword(String email, String password);
@@ -29,95 +16,84 @@ abstract class AuthBase {
 class Auth implements AuthBase {
   final _firebaseAuth = FirebaseAuth.instance;
 
-  User _userFromFirebase(FirebaseUser user) {
-    if (user == null) {
-      return null;
-    }
-    return User(
-      uid: user.uid,
-      displayName: user.displayName,
-      photoUrl: user.photoUrl,
-    );
-  }
+  @override
+  Stream<User> authStateChanges() => _firebaseAuth.authStateChanges();
 
   @override
-  Stream<User> get onAuthStateChanged {
-    return _firebaseAuth.onAuthStateChanged.map(_userFromFirebase);
-  }
-
-  @override
-  Future<User> currentUser() async {
-    final user = await _firebaseAuth.currentUser();
-    return _userFromFirebase(user);
-  }
+  User get currentUser => _firebaseAuth.currentUser;
 
   @override
   Future<User> signInAnonymously() async {
-    final authResult = await _firebaseAuth.signInAnonymously();
-    return _userFromFirebase(authResult.user);
+    final userCredential = await _firebaseAuth.signInAnonymously();
+    return userCredential.user;
   }
 
   @override
   Future<User> signInWithEmailAndPassword(String email, String password) async {
-    final authResult = await _firebaseAuth.signInWithEmailAndPassword(
+    final userCredential = await _firebaseAuth.signInWithEmailAndPassword(
         email: email, password: password);
-    return _userFromFirebase(authResult.user);
+    return userCredential.user;
   }
 
   @override
   Future<User> createUserWithEmailAndPassword(
       String email, String password) async {
-    final authResult = await _firebaseAuth.createUserWithEmailAndPassword(
+    final userCredential = await _firebaseAuth.createUserWithEmailAndPassword(
         email: email, password: password);
-    return _userFromFirebase(authResult.user);
+    return userCredential.user;
   }
 
   @override
   Future<User> signInWithGoogle() async {
-    final googleSignIn = GoogleSignIn();
-    final googleAccount = await googleSignIn.signIn();
-    if (googleAccount != null) {
-      final googleAuth = await googleAccount.authentication;
-      if (googleAuth.accessToken != null && googleAuth.idToken != null) {
-        final authResult = await _firebaseAuth.signInWithCredential(
-          GoogleAuthProvider.getCredential(
-            idToken: googleAuth.idToken,
-            accessToken: googleAuth.accessToken,
-          ),
-        );
-        return _userFromFirebase(authResult.user);
+    final GoogleSignIn googleSignIn = GoogleSignIn();
+    final GoogleSignInAccount googleUser = await googleSignIn.signIn();
+
+    if (googleUser != null) {
+      final GoogleSignInAuthentication googleAuth =
+          await googleUser.authentication;
+      if (googleAuth.idToken != null) {
+        final userCredential = await FirebaseAuth.instance
+            .signInWithCredential(GoogleAuthProvider.credential(
+          idToken: googleAuth.idToken,
+          // Note: Access token is null when running on web, so we don't check for it above
+          accessToken: googleAuth.accessToken,
+        ));
+        return userCredential.user;
       } else {
-        throw PlatformException(
-          code: 'ERROR_MISSING_GOOGLE_AUTH_TOKEN',
-          message: 'Missing Google Auth Token',
-        );
+        throw FirebaseAuthException(
+            code: 'ERROR_MISSING_GOOGLE_ID_TOKEN',
+            message: 'Missing Google ID Token');
       }
     } else {
-      throw PlatformException(
-        code: 'ERROR_ABORTED_BY_USER',
-        message: 'Sign in aborted by user',
-      );
+      throw FirebaseAuthException(
+          code: 'ERROR_ABORTED_BY_USER', message: 'Sign in aborted by user');
     }
   }
 
   @override
   Future<User> signInWithFacebook() async {
-    final facebookLogin = FacebookLogin();
-    final result = await facebookLogin.logInWithReadPermissions(
-      ['public_profile'],
-    );
-    if (result.accessToken != null) {
-      final authResult = await _firebaseAuth.signInWithCredential(
-        FacebookAuthProvider.getCredential(
-          accessToken: result.accessToken.token,
-        ),
-      );
-      return _userFromFirebase(authResult.user);
-    } else {
-      throw PlatformException(
-        code: 'ERROR_ABORTED_BY_USER',
-        message: 'Sign in aborted by user',
-      );
+    final fb = FacebookLogin();
+
+    final response = await fb.logIn(permissions: [
+      FacebookPermission.publicProfile,
+      FacebookPermission.email,
+    ]);
+    switch (response.status) {
+      case FacebookLoginStatus.Success:
+        final accessToken = response.accessToken;
+        final userCredential = await _firebaseAuth.signInWithCredential(
+            FacebookAuthProvider.credential(accessToken.token));
+        return userCredential.user;
+      case FacebookLoginStatus.Cancel:
+        throw FirebaseAuthException(
+            code: 'ERROR_ABORTED_BY_USER', message: 'Sign in aborted by user');
+      case FacebookLoginStatus.Error:
+        print('Error while log in: ${response.error}');
+        throw FirebaseAuthException(
+            code: 'ERROR_FACEBOOK_LOGIN_FAILED',
+            message: response.error.developerMessage);
+      default:
+        throw UnimplementedError();
     }
   }
 
