@@ -1,9 +1,8 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:time_tracker_flutter_course/app/sign_in/sign_in_manager.dart';
-import 'package:time_tracker_flutter_course/services/auth.dart';
 
 import 'mocks.dart';
 
@@ -31,16 +30,16 @@ void main() {
   });
 
   test('sign-in - success', () async {
-    when(mockAuth.signInAnonymously())
-        .thenAnswer((_) => Future.value(User(uid: '123')));
+    final user = MockUser.uid('123');
+    when(mockAuth.signInAnonymously()).thenAnswer((_) => Future.value(user));
     await manager.signInAnonymously();
 
     expect(isLoading.values, [true]);
   });
 
   test('sign-in - failure', () async {
-    when(mockAuth.signInAnonymously())
-        .thenThrow(PlatformException(code: 'ERROR', message: 'sign-in-failed'));
+    when(mockAuth.signInAnonymously()).thenThrow(FirebaseAuthException(
+        code: 'ERROR_WRONG_PASSWORD', message: 'Sign in failed'));
     try {
       await manager.signInAnonymously();
     } catch (e) {
